Read default attributes using OpenSSL config
Without this patch the Rake tasks prompt for attributes from rake
itself.  This leads to a number of issues:

 * Password needs to be provided for each private key
 * No default attributes for SSL certificates
 * Harder to issue SSL certificates with minor customizations

This patch addresses the problem by splitting the configuration file
into two files.  The first is for the Root CA itself which provides a
default Common Name suitable for the CA.

The second configuration file is for the SSL certificates.  The
attributes are largely the same as that of the CA but the default CN
field is read from the environment rather than from the configuration
file.

This could probably be consolidated.